% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lonlat.R
\name{add_lonlat}
\alias{add_lonlat}
\title{Add columns lon and lat from an sf object}
\usage{
add_lonlat(x, drop_geom = FALSE)
}
\arguments{
\item{x}{an sf object}

\item{drop_geom}{{FALSE} run st_drop_geometry() to return only a tibble?}
}
\value{
sf object or tibble with added columns 'lon' and 'lat'
}
\description{
I don't think there's a native option for this
}
