% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allRegressions.R
\name{all.possible.regressions}
\alias{all.possible.regressions}
\title{fit all possible regressions}
\usage{
\method{all}{possible.regressions}(dat, k)
}
\arguments{
\item{dat}{matrix or dataframe containing all variables. The response variable
must be named 'y' and the predictor variables as 'x1', 'x2', ... 'xk'.}

\item{k}{number of predictor variables}
}
\value{
dataframe with a row for each model and columns with the model formula,
df.residual, p, SSE, MSE, R2, R2.adj, AIC, AICc, and BIC.
}
\description{
This function runs a linear regression for all combinations of variables. It
calculates a number of different comparative metrics
}
\keyword{all,}
\keyword{multiple,}
\keyword{possible}
\keyword{regression,}
