% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_antilogit.R
\name{antilogit}
\alias{antilogit}
\title{antilogit function}
\usage{
antilogit(x)
}
\arguments{
\item{x}{A numeric vector, matrix, or array}
}
\value{
Object of the same dimensions as x
}
\description{
This function calculates the antilogit for the object 'x' (range -Inf to Inf),
returning the corresponding proportions (range 0 to 1).
}
\examples{
x <- runif(20, -100, 100)
antlogit(x)
}
\keyword{antilogit,}
\keyword{proportion}
