% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.R
\name{dirrf}
\alias{dirrf}
\title{dir() but with full.names=T and recursive=T}
\usage{
dirrf(...)
}
\arguments{
\item{...}{}
}
\value{
A character vector containing the names of the files in the specified directories (empty if there were no files). If a path does not exist or is not a directory or is unreadable it is skipped.
}
\description{
dir() but with full.names=T and recursive=T
}
