% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{g.band}
\alias{g.band}
\title{interpolate genus richness for each elevational band}
\usage{
g.band(sp.df, els)
}
\arguments{
\item{sp.df}{Dataframe with columns 'Genus', 'LowEl', 'HighEl'}

\item{els}{Vector of elevations}
}
\description{
This function interpolates genus richness for a sequence of elevational bands based on the range limits of a set of species.
}
\examples{
sp.df <- data.frame(Binomial=LETTERS[1:10],
                    Genus=rep(letters[1:3], times=c(2,7,1)),
                    LowEl=runif(10, 0, 1000))
sp.df$HighEl <- sp.df$LowEl + runif(10, 0, 1000)
els <- seq(0, 2000, by=100)
g.band(sp.df, els)

}
\keyword{elevation,}
\keyword{interpolate}
\keyword{richness,}
