% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_intervals.R
\name{get_intervals}
\alias{get_intervals}
\title{Summarise a dataframe to mean, medians, and CIs}
\usage{
get_intervals(df, y, type = "hdci")
}
\arguments{
\item{df}{Dataframe; if grouped, output is by group}

\item{y}{Variable to summarise}

\item{type}{Type of CI; either "hdci" or "qi", using the corresponding function from tidybayes}
}
\value{
Summarised dataframe with columns mn, med, L025, L05, L10, L25, L75, L90, L95, and L975, and either a single row or one row per group
}
\description{
Summarise a dataframe to mean, medians, and CIs
}
