% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht.R
\name{ht}
\alias{ht}
\title{Return the first and last parts of an object}
\usage{
ht(x, n = 6, ...)
}
\arguments{
\item{x}{an object}

\item{n}{an integer vector of length up to \code{dim(x)}}

\item{...}{arguments to be passed to or from other methods}
}
\value{
list with head and tail
}
\description{
Uses \code{head()} and \code{tail()} because sometimes I want to see both
}
