% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_to_gif.R
\name{img_to_gif}
\alias{img_to_gif}
\title{Create a gif from a set of images}
\usage{
img_to_gif(files, fps, out_name, optimize = T, remove_img = F)
}
\arguments{
\item{files}{Vector of file paths to images, in order}

\item{fps}{Frames per second}

\item{out_name}{File name of gif output}

\item{optimize}{Argument to magick::image_animate; If \code{TRUE}, difference 
between images is stored}

\item{remove_img}{If \code{TRUE}, remove all img files; default \code{FALSE}}
}
\value{
Nothing
}
\description{
Create a gif from a set of images
}
