% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_antilogit.R
\name{logit}
\alias{logit}
\title{logit function}
\usage{
logit(p)
}
\arguments{
\item{p}{A numeric vector, matrix, or array of proportions}
}
\value{
Object of the same dimensions as p
}
\description{
This function calculates the logit for the object 'p' (range 0 to 1),
returning the corresponding values (range -Inf to Inf).
}
\examples{
p <- runif(20, 0, 1)
logit(p)
}
\keyword{logit,}
\keyword{proportion}
