% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNewFuncs.R
\name{nf}
\alias{nf}
\title{function for writing documentation for new functions}
\usage{
nf()
}
\description{
This function loads the necessary libraries and runs document() to create the 
documentation for any newly added functions.
}
