% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairsPanels.R
\name{panel.cor}
\alias{panel.cor}
\title{calculates correlations for \code{\link{pairs}}}
\usage{
panel.cor(x, y, digits = 2, prefix = "", cex.cor, ...)
}
\description{
This function is from the help file for \code{\link{pairs}}. It calculates 
correlation coefficients to be used in the panels instead of the default 
scatterplot. The font size is scaled. (panel.upper=panel.hist).
}
