% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairsPanels.R
\name{panel.hist}
\alias{panel.hist}
\title{creates histograms for \code{\link{pairs}}}
\usage{
panel.hist(x, ...)
}
\description{
This function is from the help file for \code{\link{pairs}}. It generates a 
histogram to be used in the panels instead of the default scatterplot. This
works best on the diagonal (panel.diag=panel.hist).
}
