% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ricker.R
\name{rick.mod.stoch}
\alias{rick.mod.stoch}
\title{stochastic Ricker model for simulating population growth in a community.}
\usage{
rick.mod.stoch(nSpp, tMax, r, N, K, sd)
}
\arguments{
\item{nSpp}{The number of species to be simulated}

\item{tMax}{The number of time steps to include in the simulation}

\item{r}{Vector (length nSpp) of the average intrinsic growth rate for each species}

\item{N}{Matrix (nrow=tMax, ncol=nSpp) to store abundances; N[1,] must be initialized}

\item{K}{Vector (length nSpp) of the carrying capacity for each species}

\item{sd}{Vector (length nSpp) of the standard deviation in annual growth rate for each species}
}
\value{
Updated matrix N with simulated abundances
}
\description{
This function generates stochastic population abundances for a nSpp species.
}
\examples{
nSpp <- 3; tMax <- 20
r <- runif(3, 0.75, 1.25)
sd <- runif(3, 0.001, 2)
K <- runif(3, 100, 200)
N <- matrix(nrow=tMax, ncol=nSpp)
N[1,] <- K
N <- rick.mod.stoch(nSpp, tMax, r, N, K, sd)

}
\keyword{Ricker,}
\keyword{deterministic}
\keyword{simulation,}
