% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\title{standard error function}
\usage{
se(x)
}
\arguments{
\item{x}{A numeric vector}
}
\description{
This function calculates the standard error for a vector because R for some reason doesn't.
}
\note{
NAs are removed from the calculation
}
\examples{
se(1:10)
se(c(1:10, NA))
se(c(NA, NA, 10))
se(c(NA, NA, 10, 20))
}
\keyword{error}
\keyword{standard}
