% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{sp.band}
\alias{sp.band}
\title{interpolate species richness for each elevational band}
\usage{
sp.band(sp.df, els)
}
\arguments{
\item{sp.df}{Dataframe with columns 'Binomial', 'LowEl', 'HighEl'}

\item{els}{Vector of elevations}
}
\description{
This function interpolates species richness for a sequence of elevational bands based on the range limits of a set of species.
}
\examples{
sp.df <- data.frame(Binomial=LETTERS[1:10],
                    LowEl=runif(10, 0, 1000))
sp.df$HighEl <- sp.df$LowEl + runif(10, 0, 1000)
els <- seq(0, 2000, by=100)
sp.band(sp.df, els)

}
\keyword{elevation,}
\keyword{interpolate}
\keyword{richness,}
