% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterBand.R
\name{zone.freqmx}
\alias{zone.freqmx}
\title{zonal statistics: internal function for summarizing within zones}
\usage{
zone.freqmx(zone.r, value.r)
}
\arguments{
\item{zone.r}{zonal raster}

\item{value.r}{value raster to be summarized within zones}
}
\value{
matrix (rows = unique value, columns = zones, cells = sums)
}
\description{
This function generates a matrix summing the number of cells for each zone-value
combination based on a zonal raster and a value raster.
}
\keyword{raster,}
\keyword{zonal,}
\keyword{zone}
