% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterBand.R
\name{zone.H}
\alias{zone.H}
\title{zonal statistics: Shannon H for raster values within zones}
\usage{
zone.H(zone.r, value.r, exclude.anthro = FALSE, na.val = "na")
}
\arguments{
\item{zone.r}{zonal raster}

\item{value.r}{value raster to be summarized within zones}

\item{exclude.anthro}{Logical; should NLCD anthropogenic habitats be excluded? Defaults to FALSE}

\item{na.val}{NA value in value raster; defaults to "na"}
}
\value{
named vector (names = zones, values = Shannon H)
}
\description{
This function calculates
}
\keyword{H,}
\keyword{Shannon}
\keyword{diversity,}
\keyword{raster,}
\keyword{zonal,}
\keyword{zone}
