% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterBand.R
\name{zone.mean}
\alias{zone.mean}
\title{zonal statistics: mean of raster values within zones}
\usage{
zone.mean(zone.r, value.r, exclude.anthro = FALSE, na.val = "na")
}
\arguments{
\item{zone.r}{zonal raster}

\item{value.r}{value raster to be summarized within zones}

\item{exclude.anthro}{Logical; should NLCD anthropogenic habitats be excluded? Defaults to FALSE}

\item{na.val}{NA value in value raster; defaults to "na"}
}
\value{
named vector (names = zones, values = mean of cell values)
}
\description{
This function sums raster values within zones
}
\keyword{raster,}
\keyword{sum,}
\keyword{total,}
\keyword{zonal,}
\keyword{zone}
